refactor: move EntryPointType and SelectorAndOffset to sequencer module
